interface/requirements.txt
pyglet==1.3.0
pyserial==3.4
pygame
